style: :art: move interfaces and implementations for them
