+++ CMakeLists.txt
@@ -20,21 +20,6 @@ find_package(ament_cmake_auto REQUIRED)
 ament_auto_find_build_dependencies()
 
 ament_auto_add_library(${PROJECT_NAME}_hardware SHARED
-  src/${PROJECT_NAME}/hardware/can.cpp
-  src/${PROJECT_NAME}/hardware/headlights.cpp
-  src/${PROJECT_NAME}/hardware/imu.cpp
-  src/${PROJECT_NAME}/hardware/indicator_led.cpp
-  src/${PROJECT_NAME}/hardware/thruster_direct/esc_direct.cpp
-  src/${PROJECT_NAME}/hardware/thruster_direct/servo_direct.cpp
-
-  src/${PROJECT_NAME}/hardware_model/can_model.cpp
-  src/${PROJECT_NAME}/hardware_model/headlights_model.cpp
-  src/${PROJECT_NAME}/hardware_model/imu_model.cpp
-  src/${PROJECT_NAME}/hardware_model/indicator_led_model.cpp
-  # src/${PROJECT_NAME}/hardware_model/thruster_direct/esc_direct_model.cpp
-  # src/${PROJECT_NAME}/hardware_model/thruster_direct/servo_direct_model.cpp
-  src/${PROJECT_NAME}/hardware_model/can/vesc_model.cpp
-
   include/${PROJECT_NAME}/hardware/can.hpp
   include/${PROJECT_NAME}/hardware/headlights.hpp
   include/${PROJECT_NAME}/hardware/imu.hpp
@@ -50,8 +35,23 @@ ament_auto_add_library(${PROJECT_NAME}_hardware SHARED
   # include/${PROJECT_NAME}/hardware_model/thruster_direct/servo_direct_model.hpp
   include/${PROJECT_NAME}/hardware_model/can/vesc_model.hpp
 
-  src/${PROJECT_NAME}/util/pigpio.cpp
-  src/${PROJECT_NAME}/util/linux_can.cpp
+  src/${PROJECT_NAME}/hardware/can.cpp
+  src/${PROJECT_NAME}/hardware/headlights.cpp
+  src/${PROJECT_NAME}/hardware/imu.cpp
+  src/${PROJECT_NAME}/hardware/indicator_led.cpp
+  src/${PROJECT_NAME}/hardware/thruster_direct/esc_direct.cpp
+  src/${PROJECT_NAME}/hardware/thruster_direct/servo_direct.cpp
+
+  src/${PROJECT_NAME}/hardware_model/can_model.cpp
+  src/${PROJECT_NAME}/hardware_model/headlights_model.cpp
+  src/${PROJECT_NAME}/hardware_model/imu_model.cpp
+  src/${PROJECT_NAME}/hardware_model/indicator_led_model.cpp
+  # src/${PROJECT_NAME}/hardware_model/thruster_direct/esc_direct_model.cpp
+  # src/${PROJECT_NAME}/hardware_model/thruster_direct/servo_direct_model.cpp
+  src/${PROJECT_NAME}/hardware_model/can/vesc_model.cpp
+
+  src/${PROJECT_NAME}/hardware_model/impl/pigpio.cpp
+  src/${PROJECT_NAME}/hardware_model/impl/linux_can.cpp
 )
 
 find_library(PIGPIO_LIBRARY pigpiod_if2)
