+++ include/sinsei_umiusi_control/hardware_model/imu_model.hpp
@@ -7,8 +7,8 @@
 #include <rclcpp/rclcpp.hpp>
 #include <rcpputils/tl_expected/expected.hpp>
 
+#include "sinsei_umiusi_control/hardware_model/interface/gpio_interface.hpp"
 #include "sinsei_umiusi_control/state/imu.hpp"
-#include "sinsei_umiusi_control/util/gpio_interface.hpp"
 
 namespace sinsei_umiusi_control::hardware_model {
