+++ include/sinsei_umiusi_control/hardware_model/indicator_led_model.hpp
@@ -6,7 +6,7 @@
 #include <string>
 
 #include "sinsei_umiusi_control/cmd/indicator_led.hpp"
-#include "sinsei_umiusi_control/util/gpio_interface.hpp"
+#include "sinsei_umiusi_control/hardware_model/interface/gpio_interface.hpp"
 
 namespace sinsei_umiusi_control::hardware_model {
