+++ include/sinsei_umiusi_control/hardware_model/interface/can_interface.hpp
@@ -1,5 +1,5 @@
-#ifndef SINSEI_UMIUSI_CONTROL_UTIL_CAN_INTERFACE_HPP
-#define SINSEI_UMIUSI_CONTROL_UTIL_CAN_INTERFACE_HPP
+#ifndef SINSEI_UMIUSI_CONTROL_HARDWARE_MODEL_INTERFACE_CAN_INTERFACE_HPP
+#define SINSEI_UMIUSI_CONTROL_HARDWARE_MODEL_INTERFACE_CAN_INTERFACE_HPP
 
 #include <cstdint>
 #include <rcpputils/tl_expected/expected.hpp>
@@ -34,4 +34,4 @@ class CanInterface {
 
 }  // namespace sinsei_umiusi_control::util
 
-#endif  // SINSEI_UMIUSI_CONTROL_UTIL_CAN_INTERFACE_HPP
+#endif  // SINSEI_UMIUSI_CONTROL_HARDWARE_MODEL_INTERFACE_CAN_INTERFACE_HPP
