+++ include/sinsei_umiusi_control/hardware_model/interface/gpio_interface.hpp
@@ -1,5 +1,5 @@
-#ifndef SINSEI_UMIUSI_CONTROL_UTIL_GPIO_HPP
-#define SINSEI_UMIUSI_CONTROL_UTIL_GPIO_HPP
+#ifndef SINSEI_UMIUSI_CONTROL_HARDWARE_MODEL_INTERFACE_GPIO_INTERFACE_HPP
+#define SINSEI_UMIUSI_CONTROL_HARDWARE_MODEL_INTERFACE_GPIO_INTERFACE_HPP
 
 #include <cstddef>
 #include <cstdint>
@@ -81,4 +81,4 @@ class GpioInterface {
 
 }  // namespace sinsei_umiusi_control::util
 
-#endif  // SINSEI_UMIUSI_CONTROL_UTIL_GPIO_HPP
+#endif  // SINSEI_UMIUSI_CONTROL_HARDWARE_MODEL_INTERFACE_GPIO_INTERFACE_HPP
