+++ src/sinsei_umiusi_control/hardware/headlights.cpp
@@ -2,8 +2,8 @@
 
 #include <memory>
 
+#include "sinsei_umiusi_control/hardware_model/impl/pigpio.hpp"
 #include "sinsei_umiusi_control/util/params.hpp"
-#include "sinsei_umiusi_control/util/pigpio.hpp"
 #include "sinsei_umiusi_control/util/serialization.hpp"
 
 namespace suchw = sinsei_umiusi_control::hardware;
