+++ src/sinsei_umiusi_control/hardware/indicator_led.cpp
@@ -1,7 +1,7 @@
 #include "sinsei_umiusi_control/hardware/indicator_led.hpp"
 
+#include "sinsei_umiusi_control/hardware_model/impl/pigpio.hpp"
 #include "sinsei_umiusi_control/util/params.hpp"
-#include "sinsei_umiusi_control/util/pigpio.hpp"
 #include "sinsei_umiusi_control/util/serialization.hpp"
 
 namespace suchw = sinsei_umiusi_control::hardware;
