+++ src/sinsei_umiusi_control/hardware_model/can/vesc_model.cpp
@@ -1,7 +1,7 @@
 #include "sinsei_umiusi_control/hardware_model/can/vesc_model.hpp"
 
+#include "sinsei_umiusi_control/hardware_model/interface/can_interface.hpp"
 #include "sinsei_umiusi_control/util/byte.hpp"
-#include "sinsei_umiusi_control/util/can_interface.hpp"
 
 namespace suc = sinsei_umiusi_control;
 namespace suchm = suc::hardware_model;
